\name{calcBias}
\alias{calcBias}
\title{calculate bias from output of testPref}
\usage{
calcBias(x, lambda, gamma)
}
\arguments{
  \item{x}{a testPref object as returned by the eponymous
  function}

  \item{lambda}{true values of lambda; TxS numbers in a
  data.frame}

  \item{gamma}{true values of gamma; TxS numbers in a
  data.frame}
}
\description{
calculate bias from output of testPref
}

