% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hypotheses.R
\name{checkHypotheses}
\alias{checkHypotheses}
\title{function to check user specified hypotheses}
\usage{
checkHypotheses(hyp)
}
\arguments{
\item{hyp}{a 2-tuple specifying the null and alternative hypotheses, respectively}
}
\description{
function to check user specified hypotheses
}

