% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/estimateGen.R
\name{estGen}
\alias{estGen}
\title{estimate a reparameterization of the  hypothesis c_st}
\usage{
estGen(Xdst, Ydst, J, I, EM, em_maxiter, BALANCED)
}
\arguments{
\item{Xdst}{matrix of sums of number of eaten prey species s during occurrence t; rows indexed by time, and cols indexed by prey species, TxS}

\item{Ydst}{matrix sum of number of caught prey species s during occurrence t; rows indexed by time, and cols indexed by prey species, TxS}

\item{J}{vector of predators caught in each time period}

\item{I}{vector of number of days all traps were left out in a given time period}

\item{EM}{boolean; whether or not EM algorithm is used}

\item{em_maxiter}{integer specifying max number of EM iterations}

\item{BALANCED}{boolean; whether or not data are BALANCED}
}
\description{
estimates parameters from hypothesis lambda != gamma, where lambda
is indepdent of gamma
}
\details{
There are 2*S*T free parameters under this hypothesis.
}

