\name{format.predPref}
\alias{format.predPref}
\title{format method for predPref objects}
\usage{
\method{format}{predPref}(x, ...)
}
\arguments{
  \item{x}{a predPref object, as output by function
  predPref}

  \item{...}{additional arguments affecting the formatting}
}
\description{
format method for predPref objects
}

