% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util.R
\name{getTimeCounts}
\alias{getTimeCounts}
\title{sum specified columns by time}
\usage{
getTimeCounts(data, vars)
}
\arguments{
\item{data}{a dataframe}

\item{vars}{column variables in data to sum over}
}
\description{
sum specified columns by time
}

