% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util.R
\name{getUnitCounts}
\alias{getUnitCounts}
\title{count number of spiders or traps in each unit of time}
\usage{
getUnitCounts(data)
}
\arguments{
\item{data}{a dataframe}
}
\description{
count number of spiders or traps in each unit of time
}

