\name{plotTestPref}
\alias{plotTestPref}
\title{plot the output of testPref}
\usage{
plotTestPref(x, hypothesis = "null")
}
\arguments{
  \item{x}{a testPref object as returned by the eponymous
  function}

  \item{hypothesis}{specify which hypothesis to plot}
}
\description{
plot the output of testPref
}

