\name{predPref}
\alias{predPref}
\title{estimates parameters of predator preferences model and calculates LRT}
\usage{
predPref(eaten, caught, alpha = 0.05)
}
\arguments{
  \item{eaten}{a df of eatings preferences; TxS}

  \item{caught}{a df of caught prey species; TxS}

  \item{alpha}{LRT level of significance}
}
\description{
estimates parameters of predator preferences model and
calculates LRT
}

