% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary.predPref.R
\name{print.summary.predPref}
\alias{print.summary.predPref}
\title{predPref summary print}
\usage{
\method{print}{summary.predPref}(x, ...)
}
\arguments{
\item{x}{object of class predPref}

\item{...}{additional arguments}
}
\description{
printing method for the summary function for class predPref
}

