\name{summary.testPref}
\alias{summary.testPref}
\title{summarize output from testPref()}
\usage{
\method{summary}{testPref}(object, ..., fun = mean)
}
\arguments{
  \item{object}{a testPref object as returned by the
  eponymous function}

  \item{...}{additional arguments}

  \item{fun}{a vectorized function to summarize each column
  by; defaults to mean()}
}
\description{
summarize output from testPref()
}

