\name{calcBias}
\alias{calcBias}
\title{calculate bias from output of testPref}
\usage{
calcBias(null, alt, lambda, gamma)
}
\arguments{
  \item{null}{null hypothesis dataset}

  \item{alt}{alternative hypothesis dataset}

  \item{lambda}{true values of lambda; TxS numbers}

  \item{gamma}{true values of gamma; TxS numbers}
}
\description{
calculate bias from output of testPref
}

