\name{converged}
\alias{converged}
\title{checks convergence of the parameters for the estimation functions}
\usage{
converged(theta, theta_old, eps = 0.001)
}
\arguments{
  \item{theta}{an object, convertible to a matrix, of
  current parameter values}

  \item{theta_old}{an object, convertible to a matrix, of
  old parameter values}

  \item{eps}{tolerance to determine convergence}
}
\description{
checks convergence of the parameters for the estimation
functions
}

