\name{getTimeCounts}
\alias{getTimeCounts}
\title{sum specified columns by time}
\usage{
getTimeCounts(data, vars, by)
}
\arguments{
  \item{data}{a dataframe}

  \item{vars}{column variables in data to sum over}

  \item{by}{extra variables to sum by}
}
\description{
sum specified columns by time
}

