\name{mean.testPref}
\alias{mean.testPref}
\title{summarize output from testPref()}
\usage{
\method{mean}{testPref}(x, ..., hypothesis = c("null", "alt", "both"))
}
\arguments{
  \item{x}{a testPref object as returned by the eponymous
  function}

  \item{...}{additional arguments}

  \item{hypothesis}{specify which hypothesis}
}
\description{
summarize output from testPref()
}

