\name{mean.testPref}
\alias{mean.testPref}
\title{summarize output from testPref()}
\usage{
\method{mean}{testPref}(object, ..., hypothesis = c("null", "alt", "both"))
}
\arguments{
  \item{object}{a testPref object as returned by the
  eponymous function}

  \item{...}{additional arguments}

  \item{hypothesis}{specify which hypothesis to plot}
}
\description{
summarize output from testPref()
}

