\name{plotTestPref}
\alias{plotTestPref}
\title{plot the output of testPref}
\usage{
plotTestPref(x, hypothesis = "null", lambda = NULL, gamma = NULL)
}
\arguments{
  \item{x}{a testPref object as returned by the eponymous
  function}

  \item{hypothesis}{specify which hypothesis to plot}

  \item{lambda}{a matrix of true values of the parameter
  lambda; TxS}

  \item{gamma}{a matrix of true values of the parameter
  gamma; TxS}
}
\description{
plot the output of testPref
}

