\name{predpref}
\alias{predpref}
\title{estimates parameters of predator preferences model and calculates LRT}
\usage{
predpref(dataframe, J, I, alpha = 0.05)
}
\arguments{
  \item{dataframe}{a df of observed data}

  \item{J}{number of predators caught at each time; assumed
  constant for now}

  \item{I}{number of traps at each time; assumed constant
  for nwo}

  \item{alpha}{LRT level of significance}
}
\description{
estimates parameters of predator preferences model and
calculates LRT
}

