% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simPref.R
\name{simPref}
\alias{simPref}
\title{simulate data}
\usage{
simPref(S, T, J, I, lambda, gamma, EM = F)
}
\arguments{
\item{S}{number of prey species}

\item{T}{number of time periods}

\item{J}{scalar or vector (of length T) number of predators caught at each time}

\item{I}{scalar or vector (of length T) effective number of traps at each time}

\item{lambda}{matrix of rates at which predator eats prey species; TxS}

\item{gamma}{matrix of rates at which prey species is seen in habitat; TxS}

\item{EM}{boolean specifying test of EM algorithm}
}
\value{
A list consisting of two dataframes, eaten and caught, made
specifically for the function \code{predPref}.
}
\description{
simulate data for predator preferences model
}
\details{
Both lambda and gamma must be specified as a matrix with rows indexing
time and columns indexing the number of species.
}
\seealso{
\code{\link{predPref}}
}

