\name{summary.predPref}
\alias{summary.predPref}
\title{predPref summary}
\usage{
\method{summary}{predPref}(object, ..., sig.level = 0.05)
}
\arguments{
  \item{object}{predPref object as returned from
  predPref()}

  \item{...}{additional arguments}

  \item{sig.level}{significance level used in hypothesis
  test}
}
\description{
summary method for predPref objects as returned by the
function \code{predPref}
}

