\name{summary.predPref}
\alias{summary.predPref}
\title{summary method for predPref objects}
\usage{
\method{summary}{predPref}(object, sig.level = 0.05, digits = NULL, ...)
}
\arguments{
  \item{object}{predPref object as returned from
  predPref()}

  \item{sig.level}{significance level used in hypothesis
  test}

  \item{digits}{number of significant digits to be printed}

  \item{...}{additional arguments affecting the call to
  format within summary fn}
}
\description{
summary method for predPref objects
}

